Refactor tests and reports into individual files Additionally create a result wrapper to organize output better
